+++ src/ContextLogger/ContextLoggerAware.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace TomPHP\ContextLogger;
+
+use TomPHP\ContextLogger;
+
+interface ContextLoggerAware
+{
+    /**
+     * @return void
+     */
+    public function setLogger(ContextLogger $logger);
+}