+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2016-11-24
+### Added
+* `TomPHP\ContextLogger\ContextLoggerAware` interface.
+* `TomPHP\ContextLogger\ContextLoggerAwareTrait` trait.
+
 ## [1.0.0] - 2016-11-20
 ### Added
-* `ContextLogger::addContext(string $name, $value)`
-* `ContextLogger::removeContext(string $name)`
+* `TomPHP\ContextLogger::addContext(string $name, $value)`
+* `TomPHP\ContextLogger::removeContext(string $name)`